api/requirements.txt
fastapi==0.104.1
python-dotenv==1.0.0
python-multipart==0.0.6
